refactor: break out analyzing method to own class.
