+++ tools/serverpod_cli/lib/src/analyzer/dart/endpoints_analyzer.dart
@@ -3,25 +3,15 @@ import 'dart:io';
 import 'package:analyzer/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer/diagnostic/diagnostic.dart';
 import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:serverpod_cli/analyzer.dart';
 import 'package:serverpod_cli/src/analyzer/code_analysis_collector.dart';
+import 'package:serverpod_cli/src/analyzer/dart/definition_analyzers/method_analyzer.dart';
 import 'package:serverpod_cli/src/analyzer/dart/definition_analyzers/parameter_analyzer.dart';
-import 'package:source_span/source_span.dart';
 import 'package:path/path.dart' as p;
 
 import 'definitions.dart';
-
-const _excludedMethodNameSet = {
-  'streamOpened',
-  'streamClosed',
-  'handleStreamMessage',
-  'sendStreamMessage',
-  'setUserObject',
-  'getUserObject',
-};
 
 /// Analyzes dart files for the protocol specification.
 class EndpointsAnalyzer {
@@ -117,10 +107,9 @@ class EndpointsAnalyzer {
               var methodDefs = <MethodDefinition>[];
               var methods = element.methods;
               for (var method in methods) {
-                // Skip private methods
-                if (method.isPrivate) continue;
-                // Skip overridden methods from the Endpoint class
-                if (_excludedMethodNameSet.contains(method.name)) continue;
+                if (!MethodAnalyzer.isEndpointMethod(method)) {
+                  continue;
+                }
 
                 var parameters =
                     ParameterAnalyzer.analyze(method.parameters, collector);
@@ -129,23 +118,17 @@ class EndpointsAnalyzer {
                   continue;
                 }
 
-                if (_isEndpointMethod(parameters.required) &&
-                    _isValidReturnType(
-                      dartType: method.returnType,
-                      dartElement: method,
-                      collector: collector,
-                    )) {
-                  var methodDef = MethodDefinition(
-                    name: method.name,
-                    documentationComment: method.documentationComment,
-                    parameters: parameters.required
-                        .sublist(1), // Skip session parameter,
-                    parametersNamed: parameters.named,
-                    parametersPositional: parameters.positional,
-                    returnType: TypeDefinition.fromDartType(method.returnType),
-                  );
-                  methodDefs.add(methodDef);
+                var methodDefinition = MethodAnalyzer.analyze(
+                  method,
+                  parameters,
+                  collector,
+                );
+
+                if (methodDefinition == null) {
+                  continue;
                 }
+
+                methodDefs.add(methodDefinition);
               }
 
               var endpointDef = EndpointDefinition(
@@ -176,82 +159,5 @@ class EndpointsAnalyzer {
     }
 
     return endpointName;
-  }
-
-  bool _isEndpointMethod(List<ParameterDefinition> paramDefs) {
-    return paramDefs.isNotEmpty && paramDefs[0].type.className == 'Session';
-  }
-
-  bool _isValidReturnType({
-    required DartType dartType,
-    required Element dartElement,
-    required CodeAnalysisCollector collector,
-  }) {
-    if (!dartType.isDartAsyncFuture) {
-      collector.addError(SourceSpanSeverityException(
-        'Return type must be a Future.',
-        dartElement.span,
-      ));
-      return false;
-    }
-
-    if (dartType is! InterfaceType) {
-      collector.addError(SourceSpanSeverityException(
-        'This type is not supported as return type.',
-        dartElement.span,
-      ));
-      return false;
-    }
-
-    var typeArguments = dartType.typeArguments;
-    if (typeArguments.length != 1) {
-      collector.addError(SourceSpanSeverityException(
-        'Future must have a type defined. E.g. Future<String>.',
-        dartElement.span,
-      ));
-      return false;
-    }
-    var innerType = typeArguments[0];
-
-    if (innerType is VoidType) {
-      return true;
-    }
-
-    if (innerType is InvalidType) {
-      collector.addError(SourceSpanSeverityException(
-        'Future has an invalid return type.',
-        dartElement.span,
-      ));
-      return false;
-    }
-
-    if (innerType is DynamicType) {
-      collector.addError(SourceSpanSeverityException(
-        'Future must have a type defined. E.g. Future<String>.',
-        dartElement.span,
-      ));
-      return false;
-    }
-
-    return true;
-  }
-}
-
-extension _DartElementSourceSpan on Element {
-  SourceSpan? get span {
-    var sourceData = source?.contents.data;
-    var sourceUri = source?.uri;
-    var offset = nameOffset;
-    var length = nameLength;
-
-    if (sourceData != null && offset != 0 && length != -1) {
-      var sourceFile = SourceFile.fromString(
-        sourceData,
-        url: sourceUri,
-      );
-      return sourceFile.span(offset, offset + length);
-    } else {
-      return null;
-    }
   }
 }
